Refactor math operations to use math32 library, transition to float32, and update geometric data structures

+++ internal/simulation/ant_os_ant_interface.go
@@ -56,7 +56,7 @@ func (a *AntOS) GoForward(steps int) {
 }
 
 func (a *AntOS) Turn(direction int) {
-	fDirection := float64(direction)
+	fDirection := float32(direction)
 	if _, ok := a.State.(*AntOSMoving); !ok {
 		a.State = &AntOSMoving{
 			TargetDirection: &fDirection,
